# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для анализа композиционного центра произведений искусства",
        "title": "Art Analysis API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет JWT токен в черный список и выполняет деавторизацию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном выходе\" {\"message\": \"Деавторизация прошла успешно\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные аутентифицированного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получить текущего пользователя",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет пароль текущего аутентифицированного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Обновить данные пользователя",
                "parameters": [
                    {
                        "description": "Новые данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_UserUpd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_User"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Выполняет вход пользователя и возвращает JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_UserReg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токен и данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_TokenLogin"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Создает нового пользователя с указанными логином и паролем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_UserReg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Зарегистрированный пользователь",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_User"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.DTO_Req_UserReg": {
            "description": "Данные для регистрации нового пользователя",
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "art_user"
                },
                "password": {
                    "type": "string",
                    "example": "secure_password_123"
                }
            }
        },
        "handler.DTO_Req_UserUpd": {
            "description": "Данные для обновления пароля пользователя",
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "new_secure_password"
                }
            }
        },
...
        "handler.DTO_Resp_TokenLogin": {
            "description": "JWT токен и данные пользователя",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/handler.DTO_User"
                }
            }
        },
...
        "handler.DTO_Resp_User": {
            "description": "Упрощенные данные пользователя",
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "art_researcher"
                }
            }
        },
        "handler.DTO_User": {
            "description": "Информация о пользователе системы",
            "type": "object",
            "properties": {
                "id_user": {
                    "type": "integer",
                    "example": 1
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "login": {
                    "type": "string",
                    "example": "art_researcher"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}